Added NetworkTester
Added the network tester class to check/wait for internet access before
trying to download a picture.

+++ WallMixer/MainViewModel.cs
@@ -20,11 +20,11 @@
         private IWallpaperRepository _db;
         private IMetroDialog _dialog;
         private Random _rand;
-        private string _temp = Environment.ExpandEnvironmentVariables("%temp%");
         private string currentImage;
         private CancellationTokenSource _cts;
 
         public string ToolTipInfo { get; set; }
+        public bool OptionEnabled { get; set; }
         
         [ImportingConstructor]
         public MainViewModel(IWindowManager manager, IWallpaperRepository db, IMetroDialog dialog)
@@ -40,29 +40,45 @@
         {
             while (true)
             {
-                _cts = new CancellationTokenSource();
-                var sources = await _db.GetSourcesAsync();
-                if (sources.Count > 0)
+                if (await NetworkTester.HasInternetAccess())
                 {
-                    var randSource = sources[_rand.Next(sources.Count)];
-                    if (randSource.Source == Source.Wallhaven) randSource.WallhavenOptions = await _db.GetWallhavenOptions(randSource);
-                    string imageUrl = await UrlFetcher.GetRandomImageUrl(randSource, await _db.ImgurClientId());
-                    string tempFile = await WebImage.DownloadImage(imageUrl);
-                    var image = Image.FromFile(tempFile);
-                    DesktopBackground.SetWallpaper(tempFile);
-                    ToolTipInfo = string.Format("Source: {0}\r\nResolution: {1}",
-                        randSource.Source == Source.Reddit ? randSource.Query : randSource.Source + " / " + randSource.Query, image.Width + "x" + image.Height);
-                    NotifyOfPropertyChange(() => ToolTipInfo);
-                    image.Dispose();
-                    if (!string.IsNullOrEmpty(currentImage)) File.Delete(currentImage);
-                    currentImage = tempFile;
+                    _cts = new CancellationTokenSource();
+                    var sources = await _db.GetSourcesAsync();
+                    if (sources.Count > 0)
+                    {
+                        OptionEnabled = true;
+                        NotifyOfPropertyChange(() => OptionEnabled);
+                        var randSource = sources[_rand.Next(sources.Count)];
+                        if (randSource.Source == Source.Wallhaven) randSource.WallhavenOptions = await _db.GetWallhavenOptions(randSource);
+                        string imageUrl = await UrlFetcher.GetRandomImageUrl(randSource, await _db.ImgurClientId());
+                        string tempFile = await WebImage.DownloadImage(imageUrl);
+                        var image = Image.FromFile(tempFile);
+                        DesktopBackground.SetWallpaper(tempFile);
+                        ToolTipInfo = string.Format("Source: {0}\r\nResolution: {1}",
+                            randSource.Source == Source.Reddit ? randSource.Query : randSource.Source + " / " + randSource.Query, image.Width + "x" + image.Height);
+                        NotifyOfPropertyChange(() => ToolTipInfo);
+                        image.Dispose();
+                        if (!string.IsNullOrEmpty(currentImage)) File.Delete(currentImage);
+                        currentImage = tempFile;
+                    }
+                    else
+                    {
+                        ToolTipInfo = "Please add sources!";
+                        NotifyOfPropertyChange(() => ToolTipInfo);
+                        OptionEnabled = false;
+                        NotifyOfPropertyChange(() => OptionEnabled);
+                    }
+                    await Task.Delay(TimeSpan.FromMinutes(await _db.Interval()), _cts.Token).ContinueWith(tsk => { });
                 }
                 else
                 {
-                    ToolTipInfo = "Please add sources!";
+                    ToolTipInfo = "Waiting for internet access...";
                     NotifyOfPropertyChange(() => ToolTipInfo);
+                    OptionEnabled = false;
+                    NotifyOfPropertyChange(() => OptionEnabled);
+                    await Task.Delay(TimeSpan.FromSeconds(30));
                 }
-                await Task.Delay(TimeSpan.FromMinutes(await _db.Interval()), _cts.Token).ContinueWith(tsk => { });
+                
             }
         }
